+++ tensorflow_decision_forests/__init__.py
@@ -51,10 +51,10 @@ loaded_model = tf_keras.models.load_model("/tmp/my_saved_model")
 ```
 """
 
-__version__ = "1.11.0"
+__version__ = "1.12.0"
 __author__ = "Mathieu Guillame-Bert"
 
-compatible_tf_versions = ["2.18.0"]
+compatible_tf_versions = ["2.19.0"]
 __git_version__ = "HEAD"  # Modify for release build.
 
 from tensorflow_decision_forests.tensorflow import check_version